+++ src/Symfony/Component/Routing/CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
  * deprecated the `RouteCompiler::REGEX_DELIMITER` constant
  * added `ExpressionLanguageProvider` to expose extra functions to route conditions
  * added support for a `stateless` keyword for configuring route stateless in PHP, YAML and XML configurations.
+ * added the "hosts" option to be able to configure the host per locale.
 
 5.0.0
 -----